Moved all Areas into one file, Areas.py.
